fix(cluster): install openjdk-11 for offline install
since we bumped the jre used by scylla-jmx from openjdk-8
to openjdk-11, and we started to check for jre-11. despite
that scylla-jmx still works with jre-8, jre-11 is the
recommended JRE for running scylla-jmx, let's use openjdk-11
for testing offline installation.

please note, https://github.com/scylladb/scylla-jmx/pull/198
was also created to address this issue. but in the long
run, this fix is a better one. as it's always desirable to
use a better supported jre in testing.

Refs https://github.com/scylladb/scylla-jmx/issues/127
Fixes https://github.com/scylladb/scylladb/issues/13414
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>